+++ yggdrasil_decision_forests/port/python/ydf/deep/generic_jax.py
@@ -517,7 +517,7 @@ class GenericJAXModel(generic_model.GenericModel):
       path: str,
       input_model_signature_fn: Any = None,
       *,
-      mode: Literal["keras", "tf"] = "keras",
+      mode: Literal["keras", "tf"] = "tf",
       feature_dtypes: Dict[str, "export_tf.TFDType"] = {},  # pytype: disable=name-error
       servo_api: bool = False,
       feed_example_proto: bool = False,